ColorPanel to a new usercontrol for replication. Hacky - will change this.
