+++ TileIconifier/Controls/IconifierPanel/PictureBox/AlignForm.cs
