+++ TileIconifier/Controls/IconifierPanel/PictureBox/AlignForm.resx
