+++ TileIconifier/Controls/IconifierPanel/PictureBox/PannablePictureBoxControlPanel.cs
