+++ TileIconifier/Controls/IconifierPanel/PictureBox/PannablePictureBoxControlPanel.resx
