+++ TileIconifier/Controls/IconifierPanel/PictureBox/PannablePictureBoxImage.cs
