+++ TileIconifier/Controls/IconifierPanel/PictureBox/PannablePictureBoxMetaData.cs
