+++ TileIconifier/Controls/IconifierPanel/TileIconifierPanel.ru.resx
