+++ TileIconifier/Forms/Main/FrmBatchShortcut.cs
@@ -1,4 +1,34 @@
-﻿using System.Collections.Generic;
+﻿#region LICENCE
+
+// /*
+//         The MIT License (MIT)
+// 
+//         Copyright (c) 2016 Johnathon M
+// 
+//         Permission is hereby granted, free of charge, to any person obtaining a copy
+//         of this software and associated documentation files (the "Software"), to deal
+//         in the Software without restriction, including without limitation the rights
+//         to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//         copies of the Software, and to permit persons to whom the Software is
+//         furnished to do so, subject to the following conditions:
+// 
+//         The above copyright notice and this permission notice shall be included in
+//         all copies or substantial portions of the Software.
+// 
+//         THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//         IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//         FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//         AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//         LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//         OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+//         THE SOFTWARE.
+// 
+// */
+
+#endregion
+
+using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Forms;
 using TileIconifier.Controls.Shortcut;
@@ -8,14 +38,14 @@ namespace TileIconifier.Forms.Main
 {
     public partial class FrmBatchShortcut : SkinnableForm
     {
-        private List<ShortcutItemListViewItem> _iconifiedItems; 
+        private List<ShortcutItemListViewItem> _iconifiedItems;
 
         public FrmBatchShortcut()
         {
             InitializeComponent();
         }
 
-        private void FrmBatchShortcut_Load(object sender, System.EventArgs e)
+        private void FrmBatchShortcut_Load(object sender, EventArgs e)
         {
             GetIconifiedShortcuts();
             UpdateListViewBox();
@@ -23,14 +53,14 @@ namespace TileIconifier.Forms.Main
 
         private void GetIconifiedShortcuts()
         {
-            _iconifiedItems = ShortcutItemListViewItemLibrary.LibraryAsListViewItems.Where(s => s.ShortcutItem.IsIconified).ToList();
+            _iconifiedItems =
+                ShortcutItemListViewItemLibrary.LibraryAsListViewItems.Where(s => s.ShortcutItem.IsIconified).ToList();
         }
 
         private void UpdateListViewBox()
         {
             BuildListBoxColumns();
             UpdateListViewBoxItems();
-
         }
 
         private void BuildListBoxColumns()
@@ -52,19 +82,21 @@ namespace TileIconifier.Forms.Main
             lstIconifiedItems.SmallImageList = smallImageList;
         }
 
-        private void btnSelectAll_Click(object sender, System.EventArgs e)
+        private void btnSelectAll_Click(object sender, EventArgs e)
         {
-            foreach (ListViewItem item in lstIconifiedItems.Items)
-            {
-                item.Checked = true;
-            }
+            SetAllItemsCheckedState(true);
         }
 
-        private void btnSelectNone_Click(object sender, System.EventArgs e)
+        private void btnSelectNone_Click(object sender, EventArgs e)
+        {
+            SetAllItemsCheckedState(false);
+        }
+
+        private void SetAllItemsCheckedState(bool checkedState)
         {
             foreach (ListViewItem item in lstIconifiedItems.Items)
             {
-                item.Checked = false;
+                item.Checked = checkedState;
             }
         }
     }
