+++ TileIconifier/TileIconifier.WinForms.csproj
@@ -115,24 +115,30 @@
     <Compile Include="Controls\Eyedropper\EyedropColorPicker.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="Controls\PictureBox\AlignForm.cs" />
-    <Compile Include="Controls\PictureBox\AlignForm.Designer.cs">
+    <Compile Include="Controls\IconifierPanel\ColorPanel.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="Controls\IconifierPanel\ColorPanel.Designer.cs">
+      <DependentUpon>ColorPanel.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Controls\IconifierPanel\PictureBox\AlignForm.cs" />
+    <Compile Include="Controls\IconifierPanel\PictureBox\AlignForm.Designer.cs">
       <DependentUpon>AlignForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="Controls\PictureBox\PannablePictureBox.cs">
+    <Compile Include="Controls\IconifierPanel\PictureBox\PannablePictureBox.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="Controls\PictureBox\PannablePictureBox.Designer.cs">
+    <Compile Include="Controls\IconifierPanel\PictureBox\PannablePictureBox.Designer.cs">
       <DependentUpon>PannablePictureBox.cs</DependentUpon>
     </Compile>
-    <Compile Include="Controls\PictureBox\PannablePictureBoxControlPanel.cs">
+    <Compile Include="Controls\IconifierPanel\PictureBox\PannablePictureBoxControlPanel.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="Controls\PictureBox\PannablePictureBoxControlPanel.Designer.cs">
+    <Compile Include="Controls\IconifierPanel\PictureBox\PannablePictureBoxControlPanel.Designer.cs">
       <DependentUpon>PannablePictureBoxControlPanel.cs</DependentUpon>
     </Compile>
-    <Compile Include="Controls\PictureBox\PannablePictureBoxImage.cs" />
-    <Compile Include="Controls\PictureBox\PannablePictureBoxMetaData.cs" />
+    <Compile Include="Controls\IconifierPanel\PictureBox\PannablePictureBoxImage.cs" />
+    <Compile Include="Controls\IconifierPanel\PictureBox\PannablePictureBoxMetaData.cs" />
     <Compile Include="Controls\Shortcut\ShortcutItemListViewItemLibrary.cs" />
     <Compile Include="Controls\SortableListView.cs">
       <SubType>Component</SubType>
@@ -144,7 +150,13 @@
     <EmbeddedResource Include="Controls\AllOrCurrentUserRadios.ru.resx">
       <DependentUpon>AllOrCurrentUserRadios.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Controls\TileIconifierPanel.ru.resx">
+    <EmbeddedResource Include="Controls\IconifierPanel\ColorPanel.resx">
+      <DependentUpon>ColorPanel.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Controls\IconifierPanel\ColorPanel.ru.resx">
+      <DependentUpon>ColorPanel.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Controls\IconifierPanel\TileIconifierPanel.ru.resx">
       <DependentUpon>TileIconifierPanel.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Forms\CustomShortcutForms\FrmCustomShortcutConfirm.ru.resx">
@@ -186,10 +198,10 @@
     <EmbeddedResource Include="Properties\Strings.ru.resx">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </EmbeddedResource>
-    <Compile Include="Controls\TileIconifierPanel.cs">
+    <Compile Include="Controls\IconifierPanel\TileIconifierPanel.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="Controls\TileIconifierPanel.Designer.cs">
+    <Compile Include="Controls\IconifierPanel\TileIconifierPanel.Designer.cs">
       <DependentUpon>TileIconifierPanel.cs</DependentUpon>
     </Compile>
     <Compile Include="Forms\CustomShortcutForms\FrmCustomShortcutConfirm.cs">
@@ -283,16 +295,16 @@
     <EmbeddedResource Include="Controls\AllOrCurrentUserRadios.resx">
       <DependentUpon>AllOrCurrentUserRadios.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Controls\PictureBox\AlignForm.resx">
+    <EmbeddedResource Include="Controls\IconifierPanel\PictureBox\AlignForm.resx">
       <DependentUpon>AlignForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Controls\PictureBox\PannablePictureBox.resx">
+    <EmbeddedResource Include="Controls\IconifierPanel\PictureBox\PannablePictureBox.resx">
       <DependentUpon>PannablePictureBox.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Controls\PictureBox\PannablePictureBoxControlPanel.resx">
+    <EmbeddedResource Include="Controls\IconifierPanel\PictureBox\PannablePictureBoxControlPanel.resx">
       <DependentUpon>PannablePictureBoxControlPanel.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Controls\TileIconifierPanel.resx">
+    <EmbeddedResource Include="Controls\IconifierPanel\TileIconifierPanel.resx">
       <DependentUpon>TileIconifierPanel.cs</DependentUpon>
       <SubType>Designer</SubType>
     </EmbeddedResource>
